+++ docs/source/Changelog.rst
@@ -25,6 +25,7 @@ Added
    -  IPC2581: Exports the PCB in IPC-2581 (IPC-DPMX) format (KiCad 9)
    -  export_3d: Exports the 3D model of the PCB in BREP, GLB, STL, STEP
       and XAO (KiCad 9) (#783)
+   -  jobset: Runs a KiCad jobset (KiCad 9)
 
 -  CLI: